# openapi.yaml
openapi: 3.1.0
info:
  title: Historiador Doc API
  description: REST API for Historiador Doc — self-hosted documentation with a built-in MCP server.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
  version: 0.1.0
paths:
...
  /setup/init:
    post:
      tags:
      - setup
      operationId: init
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
        required: true
      responses:
        '200':
          description: installation initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupResponse'
        '400':
          description: validation error or LLM key rejected
        '409':
          description: setup already complete
  /setup/ollama-models:
    post:
      tags:
      - setup
      operationId: ollama_models
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OllamaModelsRequest'
        required: true
      responses:
        '200':
          description: available models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OllamaModelsResponse'
        '400':
          description: invalid URL or Ollama unreachable
  /setup/probe:
    post:
      tags:
      - setup
      operationId: probe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProbeRequest'
        required: true
      responses:
        '200':
          description: probe result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbeResponse'
components:
  schemas:
...
    LlmProvider:
      type: string
      enum:
      - openai
      - anthropic
      - ollama
      - test
...
    OllamaModelEntry:
      type: object
      required:
      - name
      - size_bytes
      properties:
        name:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 0
    OllamaModelsRequest:
      type: object
      required:
      - base_url
      properties:
        base_url:
          type: string
          description: Base URL of a reachable Ollama server (e.g. `http://localhost:11434`).
    OllamaModelsResponse:
      type: object
      required:
      - models
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/OllamaModelEntry'
...
    ProbeRequest:
      type: object
      required:
      - llm_provider
      properties:
        llm_api_key:
          type: string
        llm_provider:
          $ref: '#/components/schemas/LlmProvider'
    ProbeResponse:
      type: object
      required:
      - success
      - message
      properties:
        message:
          type: string
        success:
          type: boolean
...
    SetupRequest:
      type: object
      required:
      - admin_email
      - admin_password
      - workspace_name
      - llm_provider
      - llm_api_key
      - languages
      - primary_language
      properties:
        admin_email:
          type: string
        admin_password:
          type: string
        embedding_model:
          type:
          - string
          - 'null'
          description: Model used for chunk embeddings during publish.
        generation_model:
          type:
          - string
          - 'null'
          description: |-
            Model used for AI text generation (chat / editor). Optional for
            cloud providers (falls back to sensible defaults); required for
            Ollama because models are user-managed local pulls.
        languages:
          type: array
          items:
            type: string
        llm_api_key:
          type: string
        llm_provider:
          $ref: '#/components/schemas/LlmProvider'
        primary_language:
          type: string
        workspace_name:
          type: string
    SetupResponse:
      type: object
      required:
      - workspace_id
      - user_id
      - setup_complete
      properties:
        setup_complete:
          type: boolean
        user_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...